update space cleanup controller so that it doesn't delete sub-spaces with "active" parent space

+++ controllers/spacecleanup/space_cleanup_controller.go
@@ -92,6 +92,11 @@ func (r *Reconciler) ensureDeletionIfNeeded(logger logr.Logger, space *toolchain
 		return false, 0, nil
 	}
 
+	// check if space has an "active" parentSpace
+	if exists, err := r.parentSpaceExists(logger, space); exists {
+		return false, 0, err
+	}
+
 	timeSinceCreation := time.Since(space.GetCreationTimestamp().Time)
 	if timeSinceCreation > deletionTimeThreshold {
 		if err := r.Client.Delete(context.TODO(), space); err != nil {
@@ -106,4 +111,37 @@ func (r *Reconciler) ensureDeletionIfNeeded(logger logr.Logger, space *toolchain
 	logger.Info("Space is not ready for deletion yet", "requeue-after", requeueAfter, "created", space.CreationTimestamp)
 
 	return true, requeueAfter, nil
+}
+
+// checkParentSpace verifies if there is an "active" parentSpace for a given Space object.
+// return true/nil if there is a parentSpace
+// return true/error if something went wrong while reading parentSpace
+// return false/nil if parentSpace was not found or it's being deleted
+func (r *Reconciler) parentSpaceExists(logger logr.Logger, space *toolchainv1alpha1.Space) (bool, error) {
+	// parentSpace spec not set, returning
+	if space.Spec.ParentSpace == "" {
+		return false, nil
+	}
+
+	// search for parentSpace
+	parentSpace := &toolchainv1alpha1.Space{}
+	err := r.Client.Get(context.TODO(), types.NamespacedName{
+		Namespace: r.Namespace,
+		Name:      space.Spec.ParentSpace,
+	}, parentSpace)
+	if err != nil {
+		if errors.IsNotFound(err) {
+			logger.Info("parentSpace not found")
+			return false, nil
+		}
+		// Error reading the object - requeue the request.
+		return true, errs.Wrap(err, "unable to get the current parentSpace")
+	}
+
+	// if parent-space is already being deleted, then we can proceed with deleting the space
+	if util.IsBeingDeleted(parentSpace) {
+		logger.Info("parentSpace is being deleted..")
+		return false, nil
+	}
+	return true, nil
 }
